Fixed bug with ? in URL
Fixed a bug when arguments are sent via GET (? and & in URL).
This was causing a 404 when calling a web page with arguements. This is
because the framework try to find the document with the arguments
(impossible to find...).
The URL is now filtered and remove EVERYTHING after the first "?" but
keep the $_GET variable intact

Also began themes functions integration. This should be released in a
few weeks (or days...)