Fixed bug with ? in URL + add default 404 page
Fixed a bug when arguments are sent via GET (? and & in URL).
This was causing a 404 when calling a web page with arguments. This is
because the framework try to find the document with the arguments
(impossible to find...).
The URL is now filtered and remove EVERYTHING after the first "?" in
$INFO['web_location'].

ADD: If 404.html does not exist in webroot, show a simple 404 error
message in Apache style instead of an incomplete web page

CHANGE: Changed variable $TEMP['show_page'] to $TEMP['render_page']

INDEV: Began themes functions integration. This should be released in a
few weeks (or days...)

NOTE: Previous commits was buggy. It was cleaning the URL when no "?"
was present... causing a bug that alway show the same page