Bug fixes and improvements
This release is meant to be the very first version beta version. It may
not be fully stable but it work. Framwork version is now 1.0 BETA
Here is the full list:

NOTE: The framework is now licensed under the Apache 2.0 license. All
files has been updated.

--{ Change log }--

Fixed: Security/Privacy issue that could allow folders and files
discovery
Fixed: The wrong 404 error document may be shown if the user call /NO404
Fixed: PHP error are still showing up when errors occur even if
debugging is disabled
Fixed: 404 when accessing root of folders
Fixed: Access to content of webroot from browsers
Added: Configuration auto-test to prevent crash

--[[ Bug fixes   ]]--
Security/Privacy issue: 403 were always handled by the server. 404 are
partially handled by the framework. They are now redirected to the
framework by default using .htaccess (may not always work. It depends on
how the server react to some conditions) and show 404 if the file/folder
does not exist in webroot folder.
This issue could allow an hacker to find hidden folders and (possibly)
scripts and use the found files to execute arbitrary code if the
targeted file is an executable script or privacy trouble if the targeted
file is not properly protected (its still possible but without directory
listing, its almost impossible).

404 with no 404 document: The framework was using "/NO404" for DOCPATH
when no 404 document was present in the webroot folder. This could show
the wrong page and allow the hard-coded 404 error document to show up
when no 404 error happened. It is now using "404".

Debugging: The error_reporting setting is now set to 0 when debugging is
disabled. This prevent weird messages (for users) from showing up when
the server is running with E_ALL (or anything else) by default.

404 when accessing root of folders: When accessing the root of folders
(ex.: /framework/somefolder), a 404 is shown even if the default
document is present (defined by the config). This was caused by a
missing check when searching which file to show.

Access to content of webroot from browsers: It is now possible to
"download" files that are in the webroot folder. Check the configuration
for allowed file's extension.

--[[ Improvements ]]--
Auto-test: First implementation of the automatic configuration testing.
This auto-test verify a few settings that are critical for running and
show error with the bad settings name if something is wrong (also send
the HTTP 500 error header).
Those tests are run before the kernel is booted so it's not possible to
override them without modifying the main code or the config file.

Themes: The first functions are now being written. Currently, there is
two: "theme_basic_page" and "theme_full_page". Still need to improve how
they work (and their name).